feat: Handle milisecond and invalid input
I check if the input string is an only integer input (which means it is miliseconds) and
output the wished output json format.
If the input string does not satisfy that condition, I output the wished error json output.

From the current perspective this means 2/4 tasks handled.
Empty must will be handled next. The most complicated is handling any kind of Date string that new Date(dateString) can
handle. I'll do that as the last task.