feat: FINISH! Handle Date String as input
Last thing to do was to handle the input if it was a date string
that is accepted by date.parse().
But before that a mistake lurked in the part where I tried to make sure
if the input was only an integer, so i could treat that as a timestamp.
parseInt() turned out to be a horrible choice. It does not care about non-integers,
it ignores them simply. So you give parseInt("2014-abasv") it gives you 2014.
I was expecting it gives me error. So this needed to be fixed.
I chose the simple \D regex, which identifies any non-digit. That worked beautifully.
The rest was easy.

DONE BABY!